requirements.txt
boto3==1.9.93
botocore==1.12.93
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==1.11
django-cors-headers==2.4.0
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-filters==0.10.2.post0
docutils==0.14
idna==2.8
jmespath==0.9.3
psycopg2-binary==2.7.7
python-dateutil==2.7.5
pytz==2018.9
qrcode==6.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
